[refactor] #30 - jwt 를 도입하면서 userId 를 직접 받지 않도록 수정

+++ order/src/main/java/com/hanghae/order/application/OrderService.java
@@ -75,7 +75,7 @@ public class OrderService {
     }
 
     @Transactional(readOnly = true)
-    public List<SimpleOrderDto> getOrders(Integer userId) {
+    public List<SimpleOrderDto> getOrders(Long userId) {
         List<Order> orders = orderRepository.findByUserIdAndStatusIsNotPending(userId);
 
         return orders.stream().map(SimpleOrderDto::from).toList();
